adapt to dclare#42: rename of SetableModifier to CoreSetableModifier
