+++ runtime/src/main/java/org/modelingvalue/dclare/mps/DNode.java
@@ -15,7 +15,7 @@
 
 package org.modelingvalue.dclare.mps;
 
-import static org.modelingvalue.dclare.SetableModifier.*;
+import static org.modelingvalue.dclare.CoreSetableModifier.*;
 
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
